% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jcENOE.R
\name{jcENOE}
\alias{jcENOE}
\title{Working with datasets from the National Occupation and Employment Survey.}
\usage{
jcENOE(coe1, coe2, u)
}
\arguments{
\item{coe1}{A dataset}

\item{coe1}{A dataset}
}
\description{
It allows to work with datasets from the National Occupation and Employment Survey.
}

